Share verify flow logic
