+++ src/ageAssurance/components/NoAccessScreen.tsx
@@ -1,14 +1,13 @@
-import {useCallback, useEffect, useState} from 'react'
+import {useCallback, useEffect} from 'react'
 import {ScrollView, View} from 'react-native'
 import {useSafeAreaInsets} from 'react-native-safe-area-context'
-import type * as AgeRange from 'expo-age-range'
 import {msg} from '@lingui/core/macro'
 import {useLingui} from '@lingui/react'
 import {Trans} from '@lingui/react/macro'
 
 import {dateDiff, useGetTimeAgo} from '#/lib/hooks/useTimeAgo'
 import {useIsBirthdateUpdateAllowed} from '#/state/birthdate'
-import {useSession, useSessionApi} from '#/state/session'
+import {useSessionApi} from '#/state/session'
 import {DeactivateAccountDialog} from '#/screens/Settings/components/DeactivateAccountDialog'
 import {DeleteAccountDialog} from '#/screens/Settings/components/DeleteAccountDialog'
 import {atoms as a, useBreakpoints, useTheme, web} from '#/alf'
@@ -30,18 +29,12 @@ import * as Toast from '#/components/Toast'
 import {Text} from '#/components/Typography'
 import {BottomSheetOutlet} from '#/../modules/bottom-sheet'
 import {useAgeAssurance} from '#/ageAssurance'
-import {
-  getDeviceSignals,
-  setDeviceSignalsForRegion,
-  useAgeAssuranceServerDataContext,
-} from '#/ageAssurance/data'
-import {logger} from '#/ageAssurance/logger'
-import {unsafeGetAndComputeAgeAssurance} from '#/ageAssurance/state'
+import {useAgeAssuranceServerDataContext} from '#/ageAssurance/data'
 import {useComputeAgeAssuranceRegionAccess} from '#/ageAssurance/useComputeAgeAssuranceRegionAccess'
+import {useAgeAssuranceVerificationFlow} from '#/ageAssurance/useVerificationFlow'
 import {
   canBirthdateUpdateIncreaseAccess,
   createGeolocationString,
-  getAgeAssuranceDataFromDeviceSignals,
   isLegacyBirthdateBug,
   useAgeAssuranceRegionConfig,
 } from '#/ageAssurance/util'
@@ -380,8 +373,6 @@ function AccessSection() {
   const control = useDialogControl()
   const appealControl = Dialog.useDialogControl()
   const getTimeAgo = useGetTimeAgo()
-  const {currentAccount} = useSession()
-  const region = useAgeAssuranceRegionConfig()
 
   const aa = useAgeAssurance()
   const {status, lastInitiatedAt} = aa.state
@@ -393,108 +384,13 @@ function AccessSection() {
   const diff = lastInitiatedAt
     ? dateDiff(lastInitiatedAt, new Date(), 'down')
     : null
-  const allowsDeviceVerification = region && aa.flags.allowsDeviceVerification
-  const verifyCta = allowsDeviceVerification
-    ? _(msg`Share age data`)
-    : hasInitiated
-      ? _(msg`Verify again`)
-      : _(msg`Verify now`)
-
-  const [isVerifyingDevice, setIsVerifyingDevice] = useState(false)
-
-  const openKwsDialog = useCallback(() => {
-    control.open()
-    ax.metric('ageAssurance:initDialogOpen', {
-      hasInitiatedPreviously: hasInitiated,
-    })
-  }, [control, ax, hasInitiated])
-
-  const onPressVerify = useCallback(async () => {
-    const did = currentAccount?.did
-
-    // Just for typescript, this screen won't be shown without a logged in user
-    if (!did) return
-
-    /*
-     * In regions that permit on-device verification, try the native age API
-     * first. We tag the result with the current region (device assurance is
-     * region-bound — a TX grant only counts in TX) and, if it's sufficient,
-     * persist it client-side so the AA state recompute lifts the gate.
-     *
-     * Once the OS returns a response we stay on the device path and report the
-     * outcome via a toast (sufficient, under-age, or no usable data) rather than
-     * silently falling back — users can still opt into KWS via the inline link.
-     * We only fall through to the KWS dialog below when the device can't give us
-     * a response at all: `getDeviceSignals` handles its own errors and returns
-     * undefined (e.g. on web or failure), and there's nothing to act on without
-     * a session DID.
-     */
-    if (allowsDeviceVerification) {
-      // Show a loading state while the OS age prompt is up.
-      setIsVerifyingDevice(true)
-      let signals: AgeRange.AgeRangeResponse | undefined
-      try {
-        signals = await getDeviceSignals()
-      } finally {
-        setIsVerifyingDevice(false)
-      }
-      if (signals) {
-        const {assuredAge} = getAgeAssuranceDataFromDeviceSignals(
-          region,
-          signals,
-        )
-        if (assuredAge !== undefined) {
-          // Persist (keyed by this region) so the AA state recomputes from the
-          // cache write. Recompute here too so we can react to the outcome: a
-          // sufficient age lifts the gate (nothing more to do), but the device
-          // may report an age below the region's threshold, in which case
-          // access stays `none` and we tell the user.
-          setDeviceSignalsForRegion({did, region, signals})
-          const {state} = unsafeGetAndComputeAgeAssurance({did})
-          if (state.access === aa.Access.None) {
-            Toast.show(
-              _(
-                msg`We're sorry, but based on the data shared by your device, you are not old enough to access Bluesky.`,
-              ),
-              {type: 'info'},
-            )
-          } else if (state.access === aa.Access.Unknown) {
-            Toast.show(
-              _(
-                msg`Hmm, it seems we weren't able to compute your level of access. Please try again.`,
-              ),
-              {type: 'warning'},
-            )
-          } else {
-            Toast.show(_(msg`Thanks! You're all set.`), {
-              type: 'success',
-            })
-          }
-          return
-        }
-        // We got a device response but it carried no usable age information.
-        Toast.show(
-          _(
-            msg`Hmm, it seems your device was unable to share age information with us.`,
-          ),
-          {type: 'warning'},
-        )
-        return
-      }
-      logger.debug(
-        `onPressVerify: no device signals available (web/error/no DID), falling back to KWS`,
-      )
-    }
-
-    openKwsDialog()
-  }, [
-    region,
-    currentAccount?.did,
+  const {
+    onPressVerify,
     openKwsDialog,
+    isVerifyingDevice,
     allowsDeviceVerification,
-    aa,
-    _,
-  ])
+    verifyCta,
+  } = useAgeAssuranceVerificationFlow({initDialogControl: control})
 
   return (
     <>
